Finished part 3 and final
In this part 3, which marked final changes to the code (unless i have things to improvise),
i have applied these changes :

1. Added the second Menu for each type of bank account.
   Each menu has it's own implementation of code, used 'when' condition
   1. Check bank account balance
   2. Withdraw money
   3. Deposit money
   4. Close the app

2. Added "transfer" function to handle withdraw and deposit.
   Used 'when' condition to pass the user's menu selection into transfer function "mode" parameter.

Notes :
For the previous 2 parts and this 3rd / final part, i do the code as what instructed from the course provided by Meta.
However, to utilize and implements my knowledge of Kotlin, i'll improvise the code to make it much better.
I prefer not to go for follow-complete-done route, instead i'll go for follow-complete-improvise-done route.