Improvised the accountBalance code for each bank account type (partial)
Problem :
This code still utilize Int.random to generate =
1. User input (from range 1 to 5)
2. Account Balance (from range 1 to 1000)
3. Amount of money for deposit/withdraw (from range 1 to 1000)

So, the problem is, all of the bank account type in this code is not a "saving" account,
especially the credit account. So it's not acceptable to have accountBalance value fixed/final for only from
range 1 to 1000 (positive).
Because the core of Credit Account is essentially user owe money to the bank, and the balance should be in negative.
When user make a payment/deposit, it reduces the negative balance, eventually bringing it back to 0.

Fix =
So, for the credit account, i implemented the range for Int.random from -1000 to 0. So the account balance never go
beyond zero (0) and "simulate" the real use case of credit bank accounts where the balance is always in negative number.
In this case, i limited to -$1000 USD.

For debit and checking account, basically both of them is a "saving" account, so using Int.random()
from 0 to 1000 (positive) range is still acceptable, since user required to have amount of money
in their account balance to access their account features such as Withdraw.
Withdraw = Account Balance - Amount of money user requested to withdraw
So, account balance for debit and checking account should not be in negative number. Otherwise, user
won't be able to withdraw money from their account when accountBalance < amountWithdraw and accountBalance == 0.